Altered Constructor for InteractableAreaController
A few commits ago, a change was made to the constructor to InteractableAreaController, adding townController as a parameter to it for the sake of the subclass NoteTakingAreaController. This caused issues with other subclasses such as ConversationAreaController and GameAreaController, causing them to fail.

This commit reverts that previous commit, and instead moves the townController parameter specifically to NoteTakingAreaController.